+++ common/src/main/java/io/github/mortuusars/envelope/world/item/LetterItem.java
@@ -8,16 +8,27 @@ import io.github.mortuusars.envelope.world.item.component.seal.Seal;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
 import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.Item;
+import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
 import org.jetbrains.annotations.NotNull;
 
-public class LetterItem extends Item implements Sealable {
-    public LetterItem(Properties properties) {
-        super(properties);
+public class LetterItem extends BlockItem implements Sealable {
+    public LetterItem(Block block, Properties properties) {
+        super(block, properties);
+    }
+
+    @Override
+    public @NotNull InteractionResult useOn(UseOnContext context) {
+        if (!context.isSecondaryUseActive()) {
+            return InteractionResult.PASS;
+        }
+        return super.useOn(context);
     }
 
     @Override
@@ -25,7 +36,7 @@ public class LetterItem extends Item implements Sealable {
         ItemStack stack = player.getItemInHand(usedHand);
         stack.set(Envelope.DataComponents.LETTER_CONTENT,
               stack.getOrDefault(Envelope.DataComponents.LETTER_CONTENT, LetterContent.EMPTY).withUnfolded(true));
-        player.getCooldowns().addCooldown(this, 3);
+        player.getCooldowns().addCooldown(this, 5);
 
         if (player instanceof ServerPlayer serverPlayer) {
             Packets.sendToClient(new OpenLetterViewScreenS2CP(usedHand), serverPlayer);