+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## UNRELEASED
 
 - Added Plains, Taiga, Savanna and Desert Dovecotes.
+- Added ability to place a Letter on the wall (by holding sneak).
 - Added ability to add a 5th box to Paper Box block, which will make it a full block.
 - Falling onto a Paper Box block will absorb the fall damage and break the block, dropping the boxes.